Potential fix for code scanning alert no. 20: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using the larger type, and the result will be correctly stored in an unsigned long.

  - Cast one of the operands to unsigned long before the multiplication.
  - Update the code on line 147 to perform the multiplication using unsigned long.

+++ libopenjpeg-2.3.1/src/lib/openjpip/cachemodel_manager.c
@@ -144,7 +144,7 @@ void print_cachemodel(cachemodel_param_t cachemodel)
     }
 
     fprintf(logstream, "\t tile header and precinct packet model:\n");
-    for (i = 0; i < target->codeidx->SIZ.XTnum * target->codeidx->SIZ.YTnum; i++) {
+    for (i = 0; i < (unsigned long)target->codeidx->SIZ.XTnum * target->codeidx->SIZ.YTnum; i++) {
         fprintf(logstream, "\t  tile.%" PRIu64 "  %d\n", i, cachemodel.th_model[i]);
         for (j = 0; j < target->codeidx->SIZ.Csiz; j++) {
             fprintf(logstream, "\t   compo.%" PRIu64 ": ", j);
